requirements.txt
torch
torchvision
numpy
Pillow
scikit-image
matplotlib
tqdm
pyyaml
opencv-python
lpips
scipy
streamlit
tensorboard
